Crate sgx_libc

Modules

ocall

Structs

addrinfo

cmsghdr

dirent

dirent64

epoll_event

fd_set

hostent

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

linger

msghdr

passwd

pollfd

pthread_attr_t

sigval

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_ll

sockaddr_nl

sockaddr_storage

sockaddr_un

stat

stat64

timespec

timeval

tm

ucred

winsize

Enums

DIR

c_void
Equivalent to C's void type when used as a pointer.

Constants

AF_ALG

AF_APPLETALK

AF_ASH

AF_ATMPVC

AF_ATMSVC

AF_AX25

AF_BLUETOOTH

AF_BRIDGE

AF_CAIF

AF_CAN

AF_DECnet

AF_ECONET

AF_IEEE802154

AF_INET

AF_INET6

AF_IPX

AF_IRDA

AF_ISDN

AF_IUCV

AF_KEY

AF_LLC

AF_LOCAL

AF_NETBEUI

AF_NETLINK

AF_NETROM

AF_PACKET

AF_PHONET

AF_PPPOX

AF_RDS

AF_ROSE

AF_ROUTE

AF_RXRPC

AF_SECURITY

AF_SNA

AF_TIPC

AF_UNIX

AF_UNSPEC

AF_WANPIPE

AF_X25

AIO_ALLDONE

AIO_CANCELED

AIO_NOTCANCELED

AT_FDCWD

AT_REMOVEDIR

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

CLOCK_BOOTTIME

CLOCK_BOOTTIME_ALARM

CLOCK_MONOTONIC

CLOCK_MONOTONIC_COARSE

CLOCK_MONOTONIC_RAW

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_REALTIME_ALARM

CLOCK_REALTIME_COARSE

CLOCK_THREAD_CPUTIME_ID

DCCP_SERVICE_LIST_MAX_LEN
maximum number of services provided on the same listening port

DCCP_SOCKOPT_AVAILABLE_CCIDS

DCCP_SOCKOPT_CCID

DCCP_SOCKOPT_CCID_RX_INFO

DCCP_SOCKOPT_CCID_TX_INFO

DCCP_SOCKOPT_CHANGE_L

DCCP_SOCKOPT_CHANGE_R

DCCP_SOCKOPT_GET_CUR_MPS

DCCP_SOCKOPT_PACKET_SIZE

DCCP_SOCKOPT_QPOLICY_ID

DCCP_SOCKOPT_QPOLICY_TXQLEN

DCCP_SOCKOPT_RECV_CSCOV

DCCP_SOCKOPT_RX_CCID

DCCP_SOCKOPT_SEND_CSCOV

DCCP_SOCKOPT_SERVER_TIMEWAIT

DCCP_SOCKOPT_SERVICE

DCCP_SOCKOPT_TX_CCID

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EFD_CLOEXEC

EFD_NONBLOCK

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2NSYNC

EL2HLT

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPOLLERR

EPOLLET

EPOLLEXCLUSIVE

EPOLLHUP

EPOLLIN

EPOLLMSG

EPOLLONESHOT

EPOLLOUT

EPOLLPRI

EPOLLRDBAND

EPOLLRDHUP

EPOLLRDNORM

EPOLLWAKEUP

EPOLLWRBAND

EPOLLWRNORM

EPOLL_CLOEXEC

EPOLL_CTL_ADD

EPOLL_CTL_DEL

EPOLL_CTL_MOD

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ESGX

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

FD_CLOEXEC

FD_SETSIZE

FILENAME_MAX

FIOCLEX

FIONBIO

FOPEN_MAX

F_ADD_SEALS

F_CANCELLK

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLEASE

F_GETLK

F_GETOWN

F_GETPIPE_SZ

F_GET_SEALS

F_NOTIFY

F_OK

F_SEAL_GROW

F_SEAL_SEAL

F_SEAL_SHRINK

F_SETFD

F_SETFL

F_SETLEASE

F_SETLK

F_SETLKW

F_SETOWN

F_SETPIPE_SZ

IFF_ALLMULTI

IFF_AUTOMEDIA

IFF_BROADCAST

IFF_DEBUG

IFF_DYNAMIC

IFF_LOOPBACK

IFF_MASTER

IFF_MULTICAST

IFF_NOARP

IFF_NOTRAILERS

IFF_POINTOPOINT

IFF_PORTSEL

IFF_PROMISC

IFF_RUNNING

IFF_SLAVE

IFF_UP

IPPROTO_AH
IP6 Auth Header

IPPROTO_BEETPH

IPPROTO_COMP
IP Payload Comp. Protocol

IPPROTO_DCCP
DCCP

IPPROTO_DSTOPTS
IP6 destination option

IPPROTO_EGP
exterior gateway protocol

IPPROTO_ENCAP
encapsulation header

IPPROTO_ESP
IP6 Encap Sec. Payload

IPPROTO_FRAGMENT
IP6 fragmentation header

IPPROTO_GRE
General Routing Encap.

IPPROTO_HOPOPTS
Hop-by-hop option header

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IDP
xns idp

IPPROTO_IGMP
group mgmt protocol

IPPROTO_IP

IPPROTO_IPIP
for compatibility

IPPROTO_IPV6

IPPROTO_MAX

IPPROTO_MH

IPPROTO_MPLS

IPPROTO_MTP

IPPROTO_NONE
IP6 no next header

IPPROTO_PIM
Protocol indep. multicast

IPPROTO_PUP
pup

IPPROTO_RAW
raw IP packet

IPPROTO_ROUTING
IP6 routing header

IPPROTO_RSVP
resource reservation

IPPROTO_SCTP
SCTP

IPPROTO_TCP

IPPROTO_TP
tp-4 w/ class negotiation

IPPROTO_UDP

IPPROTO_UDPLITE

IPV6_UNICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_LOOP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IPV6_RECVPKTINFO

IPV6_PKTINFO

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_PKTINFO

IP_TRANSPARENT

IP_TTL

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_WAIT

LIO_WRITE

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

L_tmpnam

MAP_ANON

MAP_ANONYMOUS

MAP_DENYWRITE

MAP_EXECUTABLE

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_NONBLOCK

MAP_POPULATE

MAP_PRIVATE

MAP_SHARED

MAP_STACK

MSG_CMSG_CLOEXEC

MSG_CONFIRM

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOR

MSG_ERRQUEUE

MSG_FASTOPEN

MSG_FIN

MSG_MORE

MSG_NOSIGNAL

MSG_OOB

MSG_PEEK

MSG_RST

MSG_SYN

MSG_TRUNC

MSG_WAITALL

MSG_WAITFORONE

MS_ASYNC

MS_INVALIDATE

MS_SYNC

NCCS

NI_DGRAM

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSERV

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECT

O_DIRECTORY

O_DSYNC

O_EXCL

O_FSYNC

O_NDELAY

O_NOATIME

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_PATH

O_RDONLY

O_RDWR

O_RSYNC

O_SYNC

O_TRUNC

O_WRONLY

PATH_MAX

PF_ALG

PF_APPLETALK

PF_ASH

PF_ATMPVC

PF_ATMSVC

PF_AX25

PF_BLUETOOTH

PF_BRIDGE

PF_CAIF

PF_CAN

PF_DECnet

PF_ECONET

PF_IEEE802154

PF_INET

PF_INET6

PF_IPX

PF_IRDA

PF_ISDN

PF_IUCV

PF_KEY

PF_LLC

PF_LOCAL

PF_NETBEUI

PF_NETLINK

PF_NETROM

PF_PACKET

PF_PHONET

PF_PPPOX

PF_RDS

PF_ROSE

PF_ROUTE

PF_RXRPC

PF_SECURITY

PF_SNA

PF_TIPC

PF_UNIX

PF_UNSPEC

PF_WANPIPE

PF_X25

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

R_OK

SCM_TIMESTAMP

SCM_TIMESTAMPING

SCM_TIMESTAMPNS

SCM_WIFI_STATUS

SEEK_CUR

SEEK_END

SEEK_SET

SFD_CLOEXEC

SFD_NONBLOCK

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGKILL

SIGPIPE

SIGQUIT

SIGSEGV

SIGTERM

SOCK_CLOEXEC

SOCK_DCCP

SOCK_DGRAM

SOCK_NONBLOCK

SOCK_PACKET

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOL_AAL

SOL_ATM

SOL_DCCP

SOL_DECNET

SOL_ICMPV6

SOL_IP

SOL_IPV6

SOL_IRDA

SOL_LLC

SOL_NETBEUI

SOL_NETLINK

SOL_PACKET

SOL_RAW

SOL_SOCKET

SOL_TCP

SOL_TIPC

SOL_UDP

SOL_X25

SOMAXCONN

SO_ACCEPTCONN

SO_ATTACH_BPF

SO_ATTACH_FILTER

SO_BINDTODEVICE

SO_BPF_EXTENSIONS

SO_BROADCAST

SO_BSDCOMPAT

SO_BUSY_POLL

SO_DEBUG

SO_DETACH_BPF

SO_DETACH_FILTER

SO_DOMAIN

SO_DONTROUTE

SO_ERROR

SO_GET_FILTER

SO_INCOMING_CPU

SO_KEEPALIVE

SO_LINGER

SO_LOCK_FILTER

SO_MARK

SO_MAX_PACING_RATE

SO_NOFCS

SO_NO_CHECK

SO_OOBINLINE

SO_PASSCRED

SO_PASSSEC

SO_PEEK_OFF

SO_PEERCRED

SO_PEERNAME

SO_PEERSEC

SO_PRIORITY

SO_PROTOCOL

SO_RCVBUF

SO_RCVBUFFORCE

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_RXQ_OVFL

SO_SECURITY_AUTHENTICATION

SO_SECURITY_ENCRYPTION_NETWORK

SO_SECURITY_ENCRYPTION_TRANSPORT

SO_SELECT_ERR_QUEUE

SO_SNDBUF

SO_SNDBUFFORCE

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TIMESTAMPING

SO_TIMESTAMPNS

SO_TYPE

SO_WIFI_STATUS

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

SYNC_FILE_RANGE_WAIT_AFTER

SYNC_FILE_RANGE_WAIT_BEFORE

SYNC_FILE_RANGE_WRITE

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_IWGRP

S_IWOTH

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TCP_CONGESTION

TCP_COOKIE_TRANSACTIONS

TCP_CORK

TCP_DEFER_ACCEPT

TCP_FASTOPEN

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_LINGER2

TCP_MAXSEG

TCP_NODELAY

TCP_QUEUE_SEQ

TCP_QUICKACK

TCP_REPAIR

TCP_REPAIR_OPTIONS

TCP_REPAIR_QUEUE

TCP_SYNCNT

TCP_THIN_DUPACK

TCP_THIN_LINEAR_TIMEOUTS

TCP_TIMESTAMP

TCP_USER_TIMEOUT

TCP_WINDOW_CLAMP

W_OK

X_OK

_IOFBF

_IOLBF

_IONBF

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_SC_2_VERSION

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_SW_DEV

_SC_2_LOCALEDEF

_SC_2_CHAR_TERM

_SC_2_UPE

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_PBS_CHECKPOINT

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_NZERO

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STREAMS

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_UIO_MAXIOV

_SC_V6_LPBIG_OFFBIG

_SC_V7_LPBIG_OFFBIG

_SC_V6_ILP32_OFFBIG

_SC_V6_ILP32_OFF32

_SC_V6_LP64_OFF64

_SC_V7_ILP32_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_LP64_OFF64

_SC_VERSION

_SC_XBS5_LPBIG_OFFBIG

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_ILP32_OFF32

_SC_XBS5_LP64_OFF64

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_SC_XOPEN_XPG2

_SC_XOPEN_XPG3

_SC_XOPEN_XPG4

Functions

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

calloc⚠

errno
Get the last error number.

error_string⚠
Gets a detailed string description for the given error number.

free⚠

malloc⚠

malloc_usable_size⚠

memalign⚠

memchr⚠

memrchr⚠

realloc⚠

set_errno
Set the last error number.

strlen⚠

Type Definitions

blkcnt64_t

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

clockid_t

dev_t

gid_t

in_addr_t

in_port_t

ino64_t

ino_t

int8_t

int16_t

int32_t

int64_t

intmax_t

intptr_t

mode_t

nfds_t

nlink_t

off64_t

off_t

pid_t

pthread_t

ptrdiff_t

sa_family_t

size_t

socklen_t

ssize_t

suseconds_t

time_t

uid_t

uint8_t

uint16_t

uint32_t

uint64_t

uintmax_t

uintptr_t